+++ projects/ionic-angular-theme-switch/src/public_api.ts
@@ -3,3 +3,5 @@
  */
 
 export * from './lib/ionic-angular-theme-switch.service';
+export * from './lib/ionic-angular-theme-switch-toggle.component';
+export * from './lib/ionic-angular-theme-switch-toggle.module';