+++ projects/ionic-angular-theme-switch/CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Add `theme-switch-toggle` component (#4)
 - Add named themes (#3)
 
 ## [0.0.3] - 2019-01-07